+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/notifications/fcm/PushNotificationClient.java
@@ -76,7 +76,7 @@ public class PushNotificationClient {
 
     /**
      * Unregister the current device from receiving incoming push notifications.
-     * All registered handlers would be removed.
+     * All registered handlers will be removed.
      * @throws RuntimeException if push notifications failed to stop.
      */
     public void stopPushNotifications() {