+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/notifications/signaling/package-info.java
@@ -3,5 +3,5 @@
 // license information.
 //
 
-/** Package containing the classes for Signaling. Real time notification service. */
+/** Package containing the classes for Signaling. Realtime notification service. */
 package com.azure.android.communication.chat.implementation.notifications.signaling;